+++ app/src/bcsc-theme/utils/getDigitalServiceCardAccountProblem.test.ts
@@ -1,6 +1,6 @@
 import { AppError, ErrorCategory } from '@/errors'
 import { AppEventCode } from '@/events/appEventCode'
-import { getPersonCredentialAccountProblem } from './personCredentialAccountProblem'
+import { getDigitalServiceCardAccountProblem } from './getDigitalServiceCardAccountProblem'
 
 const newAppError = (): AppError =>
   new AppError(
@@ -15,78 +15,78 @@ const axiosCause = (status: number, data: unknown) =>
     response: { status, data },
   })
 
-describe('getPersonCredentialAccountProblem', () => {
+describe('getDigitalServiceCardAccountProblem', () => {
   it('returns "suspended" for a 400 with a "suspended" error_description', () => {
     const error = newAppError()
     error.cause = axiosCause(400, { error: 'unauthorized_client', error_description: 'suspended' })
 
-    expect(getPersonCredentialAccountProblem(error)).toBe('suspended')
+    expect(getDigitalServiceCardAccountProblem(error)).toBe('suspended')
   })
 
   it('returns "deactivated" for a 400 with a "deactivated" error_description', () => {
     const error = newAppError()
     error.cause = axiosCause(400, { error: 'unauthorized_client', error_description: 'deactivated' })
 
-    expect(getPersonCredentialAccountProblem(error)).toBe('deactivated')
+    expect(getDigitalServiceCardAccountProblem(error)).toBe('deactivated')
   })
 
   it('matches case-insensitively', () => {
     const error = newAppError()
     error.cause = axiosCause(400, { error: 'unauthorized_client', error_description: 'Account SUSPENDED' })
 
-    expect(getPersonCredentialAccountProblem(error)).toBe('suspended')
+    expect(getDigitalServiceCardAccountProblem(error)).toBe('suspended')
   })
 
   it('returns undefined when error is not "unauthorized_client", even if error_description mentions suspended/deactivated', () => {
     const error = newAppError()
     error.cause = axiosCause(400, { error: 'some_other_error', error_description: 'account suspended' })
 
-    expect(getPersonCredentialAccountProblem(error)).toBeUndefined()
+    expect(getDigitalServiceCardAccountProblem(error)).toBeUndefined()
   })
 
   it('returns undefined when error is missing, even if error_description mentions suspended/deactivated', () => {
     const error = newAppError()
     error.cause = axiosCause(400, { error_description: 'account deactivated' })
 
-    expect(getPersonCredentialAccountProblem(error)).toBeUndefined()
+    expect(getDigitalServiceCardAccountProblem(error)).toBeUndefined()
   })
 
   it('returns undefined for a non-400 status', () => {
     const error = newAppError()
     error.cause = axiosCause(401, { error_description: 'suspended' })
 
-    expect(getPersonCredentialAccountProblem(error)).toBeUndefined()
+    expect(getDigitalServiceCardAccountProblem(error)).toBeUndefined()
   })
 
   it('returns undefined when error_description does not mention suspended/deactivated', () => {
     const error = newAppError()
     error.cause = axiosCause(400, { error: 'unauthorized_client', error_description: 'some_other_reason' })
 
-    expect(getPersonCredentialAccountProblem(error)).toBeUndefined()
+    expect(getDigitalServiceCardAccountProblem(error)).toBeUndefined()
   })
 
   it('returns undefined when error_description is missing', () => {
     const error = newAppError()
     error.cause = axiosCause(400, { error: 'unauthorized_client' })
 
-    expect(getPersonCredentialAccountProblem(error)).toBeUndefined()
+    expect(getDigitalServiceCardAccountProblem(error)).toBeUndefined()
   })
 
   it('returns undefined when error_description is not a string', () => {
     const error = newAppError()
     error.cause = axiosCause(400, { error: 'unauthorized_client', error_description: { nested: 'suspended' } })
 
-    expect(getPersonCredentialAccountProblem(error)).toBeUndefined()
+    expect(getDigitalServiceCardAccountProblem(error)).toBeUndefined()
   })
 
   it('returns undefined when the cause is not an axios error', () => {
     const error = newAppError()
     error.cause = new Error('not axios')
 
-    expect(getPersonCredentialAccountProblem(error)).toBeUndefined()
+    expect(getDigitalServiceCardAccountProblem(error)).toBeUndefined()
   })
 
   it('returns undefined for a plain (non-AppError) error', () => {
-    expect(getPersonCredentialAccountProblem(new Error('plain error'))).toBeUndefined()
+    expect(getDigitalServiceCardAccountProblem(new Error('plain error'))).toBeUndefined()
   })
 })
