+++ app/src/services/auto-credential.ts
@@ -124,14 +124,7 @@ export class AutoCredentialMonitor implements CredentialProvisioningMonitor {
 
   /**
    * Manually starts the first configured rule's workflow, bypassing the normal
-   * "proof request references a missing cred def" trigger. Intended for the
-   * Developer settings test button so the JIT credential fetch can be exercised
-   * without waiting for a real proof request.
-   *
-   * Passes a stub proof record — safe only because no configured rule's
-   * `getInvitationUrl` reads its `proof` argument today (the DigitalServicesCard
-   * rule ignores it entirely). If a future rule derives its invitation from proof
-   * restrictions, this method will need a real proof to test that rule.
+   * "proof request references a missing cred def" trigger.
    */
   public triggerTestWorkflow(): boolean {
     if (this._workflowInProgress) {
